refactor: support mantine v7 in supabase dir
