+++ app/supabase/style.module.css
@@ -1,0 +1,7 @@
+.stack {
+  font-family: 'Helvetica';
+  margin: '0 auto';
+  width: '100%';
+  max-width: 1000;
+  padding: 20;
+}
